first_data_pipeline/requirements.txt
box-python-sdk
python-dotenv
pymongo
pandas
pymysql
boxsdk
pymysql

